fix(studio): satisfy react doctor for users page
